wth - (RMID) Change eIRB Verification Criteria
Please change the verification criteria for eIRB verification (for
updating RM record long and short title) and delete the "Withdrawn"
state from the criteria (validated_states).

Background info: withdrawn studies in eIRB is not desired to be tracked
and not maintained any more, and we don't want RMID records to be
updating according to it. When it comes back for a re-submission, it
would change to another state and then be updated.

[#146791593]

Story - (https://www.pivotaltracker.com/story/show/146791593)