wth - (RMID) Association Update Logic Needed
Background: When the RMID used by a protocol in SPARC, eIRB or coeus is updated, the associations in the RMID system should be updated accordingly.

For example, SPARC protocol 11631 used to have RMID 1926, and the users updated it to be 2221, after the update, the associated record shown in RMID system is showing SPARC protocol 11631 on both RMID records.
1).The invalid association (SPARC11631 and RMID 1926) should be removed when the updated association happens.
2). Same logic applies for eIRB and coeus.
3). Please make sure that this update is reflected in the audit trail.

[#156961557]

Story - https://www.pivotaltracker.com/story/show/156961557